refactor: Update title and README, improve invoice reconciliation
Updates the application title to "GestOs - ERP Tesouraria" and revises the README to be more concise.

Improves invoice reconciliation logic in the InvoicingModule by ensuring the correct transaction date is used and simplifying the creation of new transactions. Also refactors the PurchasingModule to correctly filter out deleted purchases and updates the monthly evolution calculation to use the correct year context.